requirements.txt
accelerate==0.15.0
aiohttp==3.8.3
aiosignal==1.3.1
alabaster==0.7.12
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
anyio==3.6.2
asgiref==3.5.2
astatine==0.3.2
astor==0.8.1
astpretty==3.0.0
astroid==2.12.12
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
autoflake==1.7.7
Babel==2.11.0
bandit==1.7.4
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==2.6.0
black==22.10.0
bs4==0.0.1
cachetools==5.2.0
cairo-lang==0.10.0
certifi==2022.9.24
cf-clearance2==0.28.3
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.30.0
cloudpickle==2.1.0
cognitive-complexity==1.3.0
coverage==6.5.0
crypto-cpp-py==1.0.4
cryptography==38.0.4
cytoolz==0.12.0
darglint==1.8.1
diffusers @ git+https://github.com/huggingface/diffusers.git@dbe071924654cdd0e05099ff7a9c7ea7e89ae084
dill==0.3.6
distlib==0.3.6
dlint==0.13.0
dnspython==2.2.1
doc8==0.11.2
docformatter==1.5.0
docutils==0.17.1
domdf-python-tools==3.5.0
ecdsa==0.18.0
eradicate==2.1.0
et-xmlfile==1.1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
exceptiongroup==1.0.2
fastecdsa==2.2.3
filelock==3.8.0
flake8==4.0.1
flake8-2020==1.6.1
flake8-aaa==0.12.2
flake8-annotations==2.9.1
flake8-annotations-complexity==0.0.7
flake8-annotations-coverage==0.0.6
flake8-bandit==3.0.0
flake8-black==0.3.3
flake8-blind-except==0.2.1
flake8-breakpoint==1.1.0
flake8-broken-line==0.4.0
flake8-bugbear==22.10.27
flake8-builtins==1.5.3
flake8-class-attributes-order==0.1.3
flake8-coding==1.3.2
flake8-cognitive-complexity==0.1.0
flake8-commas==2.1.0
flake8-comments==0.1.2
flake8-comprehensions==3.10.1
flake8-debugger==4.1.2
flake8-django==1.1.5
flake8-docstrings==1.6.0
flake8-encodings==0.5.0.post1
flake8-eradicate==1.4.0
flake8-executable==2.1.2
flake8-expression-complexity==0.0.11
flake8-fixme==1.1.1
flake8-functions==0.0.7
flake8-functions-names==0.3.0
flake8-future-annotations==0.0.5
flake8-helper==0.2.1
flake8-isort==4.2.0
flake8-literal==1.2.6
flake8-logging-format==0.8.1
flake8-markdown==0.3.0
flake8-mutable==1.2.0
flake8-no-pep420==2.3.0
flake8-noqa==1.2.9
flake8-pie==0.16.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-pyi==22.10.0
flake8-pylint==0.1.3
flake8-pytest-style==1.6.0
flake8-quotes==3.3.1
flake8-rst-docstrings==0.2.7
flake8-secure-coding-standard==1.3.0
flake8-simplify==0.19.3
flake8-slots==0.1.5
flake8-string-format==0.3.0
flake8-tidy-imports==4.8.0
flake8-typing-imports==1.12.0
flake8-use-fstring==1.4
flake8-use-pathlib==0.3.0
flake8-useless-assert==0.4.3
flake8-variables-names==0.0.5
flake8-warnings==0.4.0
Flask==2.0.3
Flask-Cors==3.0.10
frozendict==1.2
frozenlist==1.3.3
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
gitdb==4.0.9
GitPython==3.1.29
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
hexbytes==0.3.0
httpcore==0.16.2
httpx==0.23.1
huggingface-hub==0.11.1
hypothesis==6.57.1
hypothesmith==0.1.9
idna==3.4
imagesize==1.4.1
importlib-metadata==5.0.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.0
jupyterlab-flake8==0.7.1
jwt==1.3.1
kucoin-futures-python==1.0.6
lark-parser==0.12.0
lazy-object-proxy==1.8.0
libcst==0.4.9
lru-dict==1.1.8
lxml==4.9.1
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.1
marshmallow==3.17.1
marshmallow-dataclass==8.5.10
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
mccabe==0.6.1
mpmath==1.2.1
mr-proper==0.0.7
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
natsort==8.2.0
nest-asyncio==1.5.6
netaddr==0.8.0
numpy==1.23.4
openai==0.25.0
OpenAIAuth==0.0.6
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
pandas-stubs==1.5.2.221124
pandas-vet==0.2.3
parsimonious==0.8.1
Paste==3.5.2
pathspec==0.9.0
pbr==5.11.0
pep8-naming==0.12.1
Pillow==9.3.0
pipdeptree==2.3.3
platformdirs==2.5.4
playwright==1.28.0
pluggy==1.0.0
prometheus-client==0.15.0
protobuf==3.19.5
psutil==5.9.4
py==1.11.0
pybetter==0.4.1
pycln==1.3.5
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
pydocstyle==6.1.1
pyee==9.0.4
pyemojify==0.2.0
pyflakes==2.4.0
Pygments==2.13.0
pylint==2.15.5
pymongo==4.3.3
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
pytest-asyncio==0.20.2
pytest-cov==3.0.0
pytest-sugar==0.9.6
python-dateutil==2.8.2
python-dev-tools==2022.5.27
python-dotenv==0.21.0
pytz==2022.6
pyupgrade==2.38.4
PyYAML==6.0
regex==2022.10.31
removestar==1.3.1
replicate==0.4.0
requests==2.28.1
restructuredtext-lint==1.4.0
revChatGPT==0.0.43.2
rfc3986==1.5.0
rlp==2.0.1
scipy==1.9.3
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.9.0
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
ssort==0.11.6
starknet-devnet @ git+https://github.com/JorikSchellekens/starknet-devnet@d226b90589e3e4f650877c65b6f1c91a28c9bda5
stdlib-list==0.8.0
stevedore==4.1.1
sympy==1.11.1
TA-Lib==0.4.25
termcolor==2.1.0
tls-client==0.1.5
tokenize-rt==4.2.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.0
torch==1.13.0
tox==3.27.1
tox-travis==0.12
tqdm==4.64.1
transformers==4.25.1
typeguard==2.13.3
typer==0.4.2
types-pytz==2022.6.0.1
typing-inspect==0.8.0
typing_extensions==4.3.0
untokenize==0.1.1
urllib3==1.26.12
varint==1.0.2
virtualenv==20.16.7
waitress==2.1.2
web3==5.31.1
websockets==9.1
wemake-python-styleguide==0.16.1
Werkzeug==2.0.3
wrapt==1.14.1
yarl==1.8.1
zipp==3.10.0